[android] migration to dedicated file for torrent preferences
